Move SchemaBuilder back to core project
