+++ src/GraphQL.Authorization/AuthorizationPolicy.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace GraphQL.Tools
+namespace GraphQL.Authorization
 {
     public interface IAuthorizationPolicy
     {
