+++ src/GraphQL.Authorization/AuthorizationPolicyBuilder.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace GraphQL.Tools
+namespace GraphQL.Authorization
 {
     public class AuthorizationPolicyBuilder
     {
