+++ src/GraphQL.Authorization/AuthorizationResult.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace GraphQL.Tools
+namespace GraphQL.Authorization
 {
     public class AuthorizationResult
     {
