+++ src/GraphQL.Authorization/ClaimAuthorizationRequirement.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace GraphQL.Tools
+namespace GraphQL.Authorization
 {
     public class ClaimAuthorizationRequirement : IAuthorizationRequirement
     {
