+++ src/GraphQL.Authorization/GraphQL.Authorization.csproj
@@ -1,18 +1,18 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <Description>GraphQL Tools for .NET</Description>
+    <Description>Authorization extensions for GraphQL .NET</Description>
     <VersionPrefix>1.0.0</VersionPrefix>
     <Authors>Joe McBride</Authors>
     <TargetFrameworks>netstandard1.3;net46</TargetFrameworks>
     <PackageTargetFallback>dnxcore50</PackageTargetFallback>
-    <AssemblyName>GraphQL.Tools</AssemblyName>
-    <PackageId>GraphQL.Tools</PackageId>
+    <AssemblyName>GraphQL.Authorization</AssemblyName>
+    <PackageId>GraphQL.Authorization</PackageId>
     <PackageTags>GraphQL;json;api</PackageTags>
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\GraphQL\GraphQL.csproj" />
+    <PackageReference Include="GraphQL" Version="0.17.1" />
     <PackageReference Include="GraphQL-Parser" Version="2.0.0" />
     <PackageReference Include="System.Security.Claims" Version="4.3.0" />
   </ItemGroup>
