+++ src/GraphQL/Utilities/TypeConfig.cs
@@ -1,54 +1,9 @@
 ﻿using System;
-using System.Collections.Concurrent;
-using System.Collections.Generic;
 using GraphQL.Resolvers;
 using GraphQL.Types;
 
-namespace GraphQL.Tools
+namespace GraphQL.Utilities
 {
-    public class MetadataProvider : IProvideMetadata
-    {
-        public IDictionary<string, object> Metadata { get; set; } = new ConcurrentDictionary<string, object>();
-
-        public TType GetMetadata<TType>(string key, TType defaultValue = default(TType))
-        {
-            if (!HasMetadata(key))
-            {
-                return defaultValue;
-            }
-
-            object item;
-            if (Metadata.TryGetValue(key, out item))
-            {
-                return (TType) item;
-            }
-
-            return defaultValue;
-        }
-
-        public bool HasMetadata(string key)
-        {
-            return Metadata.ContainsKey(key);
-        }
-    }
-
-    public class FieldConfig : MetadataProvider
-    {
-        public string Name { get; set; }
-        public string Description { get; set; }
-        public IFieldResolver ResolverValue { get; set; }
-
-        public void Resolver<TSourceType, TReturnType>(Func<ResolveFieldContext<TSourceType>, TReturnType> resolver)
-        {
-            ResolverValue = new FuncFieldResolver<TSourceType, TReturnType>(resolver);
-        }
-
-        public void Resolver<TReturnType>(Func<ResolveFieldContext, TReturnType> resolver)
-        {
-            ResolverValue = new FuncFieldResolver<TReturnType>(resolver);
-        }
-    }
-
     public class TypeConfig : MetadataProvider
     {
         private readonly LightweightCache<string, FieldConfig> _resolvers;
